+++ Sources/peertalk/include/PTPrivate.h
