+++ Tests/PeertalkTests/PTProtocolTests.h
@@ -1,7 +1,8 @@
 #import <XCTest/XCTest.h>
 #include <dispatch/dispatch.h>
-#import "PTProtocol.h"
-#import "PTPrivate.h"
+
+#import "include/PTProtocol.h"
+#import "include/PTPrivate.h"
 
 @interface PTProtocolTests : XCTestCase {
   dispatch_fd_t socket_[2];
